# hillshade: implement SSE2 optimization on Float32 input data type