This also unifies and simplifies with the existing SSE2 optimization that worked on integer data type.

On https://rockyweb.usgs.gov/vdelivery/Datasets/Staged/Elevation/1m/Projects/WA_FEMAHQ_2018_D18/TIFF/USGS_1M_10_x56y512_WA_FEMAHQ_2018_D18.tif

master:
```
$ hyperfine --warmup 2 --  "time gdaldem hillshade USGS_1M_10_x56y512_WA_FEMAHQ_2018_D18.tif mem -of mem -q"
Benchmark #1: time gdaldem hillshade USGS_1M_10_x56y512_WA_FEMAHQ_2018_D18.tif mem -of mem -q
  Time (mean ± σ):      4.379 s ±  0.257 s    [User: 4.186 s, System: 0.190 s]
  Range (min … max):    4.079 s …  4.696 s
```

this PR:
```
$ hyperfine --warmup 2 --  "time gdaldem hillshade USGS_1M_10_x56y512_WA_FEMAHQ_2018_D18.tif mem -of mem -q"
Benchmark #1: time gdaldem hillshade USGS_1M_10_x56y512_WA_FEMAHQ_2018_D18.tif mem -of mem -q
  Time (mean ± σ):      2.425 s ±  0.106 s    [User: 2.252 s, System: 0.172 s]
  Range (min … max):    2.221 s …  2.501 s
```